html/rendering/non-replaced-elements/the-fieldset-element-0/fieldset-grid.html

<!DOCTYPE html>
<title>fieldset and CSS Grid</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<style>
 #test, #ref { display: grid; grid-template-columns: auto 50px auto; grid-template-rows: auto 50px auto; margin: 0; padding: 0; border: none }
</style>
<fieldset id=test>
 <div>1</div>
 <div>2</div>
 <div>3</div>
 <div>4</div>
 <div>5</div>
 <div>6</div>
 <div>7</div>
 <div>8</div>
 <div>9</div>
</fieldset>
<hr>
<div id=ref>
 <div>1</div>
 <div>2</div>
 <div>3</div>
 <div>4</div>
 <div>5</div>
 <div>6</div>
 <div>7</div>
 <div>8</div>
 <div>9</div>
</div>
<script>
  test(() => {
    assert_equals(getComputedStyle(document.getElementById('test')).height,
                  getComputedStyle(document.getElementById('ref')).height);
  });
</script>
